TJFE-58 [design] MVI 패턴 적용 & 파일 이동(-> log_in.view)
